Add generator for running test data sets to BrotatoTestBase
